atom_chat/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.2.1
redis==5.2.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.3.0
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
websockets==14.0
zope.interface==7.1.1
